int->long, paging improvements

+++ webapp/src/main/java/si/merljak/magistrska/client/handler/PagingHandler.java
@@ -12,5 +12,5 @@ public interface PagingHandler {
 	 * 
 	 * @param page new page
 	 */
-	void onPageChange(int page);
+	void onPageChange(long page);
 }
